api/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
joblib==1.3.1
python-multipart==0.0.6
psycopg2-binary==2.9.7